Forms from views.py to forms.py
