+++ WebCLI/views/views.py
@@ -1,17 +1,13 @@
 from WebMark.settings import ALGORITHMS_PER_PAGE, ROOT_DIR
 from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
-from django.contrib.auth.forms import UserCreationForm
 from django.core.exceptions import PermissionDenied
-from django.urls import reverse_lazy
 from django.utils.html import format_html
-from django.views import generic
-from django.forms import ModelForm, Textarea, HiddenInput, Form
-from django.forms import CharField
-from django.forms.widgets import NumberInput
 from django_tables2.columns.base import Column
 from django_tables2.columns import TemplateColumn
 from ..models import Algorithm, Molecule, Algorithm_type, Algorithm_version, Metrics
+from .forms import AlgorithmForm, AlgorithmTypeForm
+from .forms import MoleculeForm, AlgorithmVersionForm, MetricsForm
 from django.utils import timezone
 from django.utils.decorators import method_decorator
 from django_filters import AllValuesFilter, FilterSet
@@ -75,41 +71,6 @@ class MyAlgorithmListView(AlgorithmListView):
         return Algorithm.objects.filter(user=self.request.user).order_by("name")
 
 
-class SignUpView(generic.CreateView):
-    form_class = UserCreationForm
-    success_url = reverse_lazy('login')
-    template_name = 'registration/signup.html'
-
-
-class AlgorithmForm(ModelForm):
-    class Meta:
-        model = Algorithm
-        fields = ['user', 'name', 'algorithm_type', 'public',
-                  'article_link', 'github_link']
-        widgets = {
-            'name': Textarea(attrs={'rows': 1, 'cols': 50}),
-            'user': HiddenInput(),
-        }
-
-
-class AlgorithmVersionForm(Form):
-    algorithm = CharField(widget=Textarea)
-
-
-class MetricsForm(ModelForm):
-    class Meta:
-        model = Metrics
-        fields = ['algorithm_version', 'molecule', 'iterations',
-                  'measurements', 'circuit_depth', 'accuracy']
-        widgets = {
-            'algorithm_version': HiddenInput(),
-            'iterations': NumberInput(attrs={'min': 0, 'max': 1000000}),
-            'measurements': NumberInput(attrs={'min': 0, 'max': 1000000}),
-            'circuit_depth': NumberInput(attrs={'min': 0, 'max': 1000000}),
-            'accuracy': NumberInput(attrs={'min': 0, 'max': 1000000}),
-        }
-
-
 @login_required
 def new_algorithm(request):
     aform = AlgorithmForm(initial={'user': request.user})
@@ -128,15 +89,6 @@ def new_algorithm(request):
     return render(request, 'WebCLI/newAlgorithm.html', data)
 
 
-class MoleculeForm(ModelForm):
-    class Meta:
-        model = Molecule
-        fields = ['name', 'structure']
-        widgets = {
-            'name': Textarea(attrs={'rows': 1, 'cols': 50}),
-        }
-
-
 @login_required
 def new_molecule(request):
     form = MoleculeForm()
@@ -145,15 +97,6 @@ def new_molecule(request):
         m.save()
     data = {'molecules': Molecule.objects.all(), 'form': form}
     return render(request, 'WebCLI/newMolecule.html', data)
-
-
-class AlgorithmTypeForm(ModelForm):
-    class Meta:
-        model = Algorithm_type
-        fields = ['type_name']
-        widgets = {
-            'type_name': Textarea(attrs={'rows': 1, 'cols': 50}),
-        }
 
 
 @login_required
